make rpc call ChangeLogLevel ctx cancellable; fix examples + add comments; fix rpc server so it is aware of both loggers;
