+++ examples/multi_service_rpc/hello_world_service.go
@@ -10,6 +10,9 @@ import (
 	"github.com/ambitiousfew/rxd/log"
 )
 
+// Ensure we meet the interface or error.
+var _ rxd.ServiceRunner = (*HelloWorldAPIService)(nil)
+
 type HelloWorldAPIService struct {
 	// fields this specific server uses
 	server *http.Server
@@ -23,16 +26,16 @@ func NewHelloWorldService() *HelloWorldAPIService {
 }
 
 // Idle can be used for some pre-run checks or used to have run fallback to an idle retry state.
-func (s *HelloWorldAPIService) Idle(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "entered idle state")
+func (s *HelloWorldAPIService) Idle(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered idle state")
 	// if all is well here, move to the RunState or retry back to Init if something went wrong.
 	timer := time.NewTimer(5 * time.Second)
 	defer timer.Stop()
 
-	// ctx.Log(log.LevelInfo, "intentionally delaying for 8s before run begi)
+	// sctx.Log(log.LevelInfo, "intentionally delaying for 8s before run begi)
 	for {
 		select {
-		case <-ctx.Done():
+		case <-sctx.Done():
 			return nil
 		case <-timer.C:
 			// Intentional 20s delay so polling service can react to failed attempts to this API.
@@ -43,27 +46,27 @@ func (s *HelloWorldAPIService) Idle(ctx rxd.ServiceContext) error {
 
 // Run is where you want the main logic of your service to run
 // when things have been initialized and are ready, this runs the heart of your service.
-func (s *HelloWorldAPIService) Run(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "entered run state")
+func (s *HelloWorldAPIService) Run(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered run state")
 	doneC := make(chan struct{})
 	go func() {
 		defer close(doneC)
 		// We should always watch for this signal, must use goroutine here
 		// since ListenAndServe will block and we need a way to end the
 		// server as well as inform the server to stop all requests ASAP.
-		<-ctx.Done()
-		ctx.Log(log.LevelInfo, "received a shutdown signal, cancel server context to stop server gracefully")
+		<-sctx.Done()
+		sctx.Log(log.LevelWarning, "received a shutdown signal, cancel server context to stop server gracefully")
 
 		timedCtx, timedCancel := context.WithTimeout(context.Background(), 3*time.Second)
 		defer timedCancel()
 
 		err := s.server.Shutdown(timedCtx)
 		if err != nil {
-			ctx.Log(log.LevelError, err.Error())
+			sctx.Log(log.LevelError, err.Error())
 		}
 	}()
 
-	ctx.Log(log.LevelInfo, fmt.Sprintf("server starting at %s", s.server.Addr))
+	sctx.Log(log.LevelInfo, fmt.Sprintf("server starting at %s", s.server.Addr))
 	// ListenAndServe will block forever serving requests/responses
 	err := s.server.ListenAndServe()
 
@@ -73,15 +76,15 @@ func (s *HelloWorldAPIService) Run(ctx rxd.ServiceContext) error {
 	}
 
 	<-doneC
-	ctx.Log(log.LevelInfo, "server shutdown")
+	sctx.Log(log.LevelInfo, "server shutdown")
 
 	// If we reached this point, we stopped the server without erroring, we are likely trying to stop our daemon.
 	// Lets stop this service properly
 	return nil
 }
 
-func (s *HelloWorldAPIService) Init(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "initializing")
+func (s *HelloWorldAPIService) Init(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered init state")
 
 	mux := http.NewServeMux()
 	mux.HandleFunc("/api", func(w http.ResponseWriter, r *http.Request) {
@@ -98,13 +101,16 @@ func (s *HelloWorldAPIService) Init(ctx rxd.ServiceContext) error {
 }
 
 // Stop handles anything you might need to do to clean up before ending your service.
-func (s *HelloWorldAPIService) Stop(ctx rxd.ServiceContext) error {
+func (s *HelloWorldAPIService) Stop(sctx rxd.ServiceContext) error {
 	// We must return a NewResponse, we use NoopState because it exits with no operation.
 	// using StopState would try to recall Stop again.
-	ctx.Log(log.LevelInfo, "stopping")
+	sctx.Log(log.LevelDebug, "entered init state")
 	s.server = nil
-	return nil
-}
 
-// Ensure we meet the interface or error.
-var _ rxd.ServiceRunner = &HelloWorldAPIService{}
+	select {
+	case <-sctx.Done():
+		return nil
+	default:
+		return nil
+	}
+}
