+++ examples/multi_service_rpc/polling_service.go
@@ -12,6 +12,9 @@ import (
 	"github.com/ambitiousfew/rxd/log"
 )
 
+// Ensure we meet the interface or error.
+var _ rxd.ServiceRunner = (*APIPollingService)(nil)
+
 // APIPollingService create a struct for your service which requires a config field along with any other state
 // your service might need to maintain throughout the life of the service.
 type APIPollingService struct {
@@ -36,18 +39,18 @@ func NewAPIPollingService() *APIPollingService {
 }
 
 // Idle can be used for some pre-run checks or used to have run fallback to an idle retry state.
-func (s *APIPollingService) Idle(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "entered idle state")
+func (s *APIPollingService) Idle(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered idle state")
 
-	statesC, cancel := ctx.WatchAllServices(rxd.Entering, rxd.StateRun, ServiceHelloWorldAPI)
+	statesC, cancel := sctx.WatchAllServices(rxd.Entering, rxd.StateRun, ServiceHelloWorldAPI)
 	defer cancel()
 
 	for {
 		select {
-		case <-ctx.Done():
+		case <-sctx.Done():
 			return nil
 		case <-statesC:
-			ctx.Log(log.LevelInfo, "Hello World API has entered Run state")
+			sctx.Log(log.LevelDebug, "leaving the entered Idle state")
 			// Hello World API should have entered a RunState, so we can move to our next state (Run) now.
 			return nil
 		}
@@ -56,23 +59,23 @@ func (s *APIPollingService) Idle(ctx rxd.ServiceContext) error {
 
 // Run is where you want the main logic of your service to run
 // when things have been initialized and are ready, this runs the heart of your service.
-func (s *APIPollingService) Run(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "entered run state")
+func (s *APIPollingService) Run(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered run state")
 	timer := time.NewTimer(1 * time.Second)
 	defer timer.Stop()
 
 	// Here we are registering our interest in ANY of the services passed EXITING a "RunState"
 	// So if any service given here for some reasons LEAVES their RunState, we will be notified.
 
-	statesC, cancel := ctx.WatchAllServices(rxd.Exiting, rxd.StateRun, ServiceHelloWorldAPI)
+	statesC, cancel := sctx.WatchAllServices(rxd.Exiting, rxd.StateRun, ServiceHelloWorldAPI)
 	defer cancel()
 
-	ctx.Log(log.LevelInfo, "starting to poll")
+	sctx.Log(log.LevelInfo, "starting to poll")
 
 	var pollCount int
 	for {
 		select {
-		case <-ctx.Done():
+		case <-sctx.Done():
 			return nil
 		case <-statesC:
 			// Hello World API should have exited Run state, so we need to move out of run too.
@@ -80,7 +83,7 @@ func (s *APIPollingService) Run(ctx rxd.ServiceContext) error {
 
 		case <-timer.C:
 			if pollCount > s.maxPollCount {
-				ctx.Log(log.LevelInfo, "reached maximum poll count, stopping")
+				sctx.Log(log.LevelInfo, "reached maximum poll count, stopping")
 				return nil
 			}
 
@@ -90,6 +93,8 @@ func (s *APIPollingService) Run(ctx rxd.ServiceContext) error {
 				timer.Reset(s.retryDuration)
 				continue
 			}
+
+			sctx.Log(log.LevelInfo, fmt.Sprintf("received response from the API: %d", resp.StatusCode))
 
 			respBytes, err := io.ReadAll(resp.Body)
 			resp.Body.Close()
@@ -106,7 +111,7 @@ func (s *APIPollingService) Run(ctx rxd.ServiceContext) error {
 				// we could return to new state: idle or stop or just continue to keep trying.
 			}
 
-			ctx.Log(log.LevelInfo, fmt.Sprintf("received response from the API: %v", respBody))
+			sctx.Log(log.LevelDebug, fmt.Sprintf("response body: %s", respBytes))
 			// Increment polling counter
 			pollCount++
 
@@ -117,17 +122,24 @@ func (s *APIPollingService) Run(ctx rxd.ServiceContext) error {
 }
 
 // Stop handles anything you might need to do to clean up before ending your service.
-func (s *APIPollingService) Stop(ctx rxd.ServiceContext) error {
+func (s *APIPollingService) Stop(sctx rxd.ServiceContext) error {
 	// We must return a NewResponse, we use NoopState because it exits with no operation.
 	// using StopState would try to recall Stop again.
-	ctx.Log(log.LevelInfo, "stopping")
-	return nil
+	sctx.Log(log.LevelDebug, "entered stop state")
+	select {
+	case <-sctx.Done():
+		return nil
+	default:
+		return nil
+	}
 }
 
-func (s *APIPollingService) Init(ctx rxd.ServiceContext) error {
-	ctx.Log(log.LevelInfo, "initializing")
-	return nil
+func (s *APIPollingService) Init(sctx rxd.ServiceContext) error {
+	sctx.Log(log.LevelDebug, "entered init state")
+	select {
+	case <-sctx.Done():
+		return nil
+	default:
+		return nil
+	}
 }
-
-// Ensure we meet the interface or error.
-var _ rxd.ServiceRunner = &APIPollingService{}
