--- pkg/rxrpc/commands.go
@@ -1,17 +1,0 @@
-package rxrpc
-
-const (
-	Unknown Command = iota
-	SetLevel
-)
-
-type Command uint8
-
-func (c Command) String() string {
-	switch c {
-	case SetLevel:
-		return "SetLevel"
-	default:
-		return "Unknown"
-	}
-}
